# Projet_Examen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portail Scolaire API",
        "description": "Documentation de l'API du portail scolaire",
        "contact": {
            "email": "diabydaj@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/enseignants": {
            "get": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Liste des enseignants",
                "operationId": "2de505ec5b8936d58e885526570f9d1e",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Créer un nouvel enseignant",
                "operationId": "b63649a22e0c435b701cc16395785b1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "email",
                                    "password",
                                    "specialite",
                                    "date_embauche",
                                    "numero_identifiant"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enseignant créé"
                    }
                }
            }
        },
        "/api/enseignants/{id}": {
            "get": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Afficher un enseignant spécifique",
                "operationId": "3604d0b5863854d2f3aae0b30c5e1b63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Enseignant non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Mettre à jour un enseignant",
                "operationId": "62ad537c342dc33310f166b825516722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "email",
                                    "specialite",
                                    "date_embauche",
                                    "numero_identifiant"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enseignant mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Supprimer un enseignant",
                "operationId": "712b167e8ecca6a82e354789b7e9bb3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Enseignant supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}